Typed notional actions and reducers
