+++ src/client/src/ui/spotTile/notional/index.ts
@@ -1,0 +1,1 @@
+export { notionalsReducer } from './reducer'
